PG-339: Fix duplicated error messages on error view.
The function pgsm_log() declares the error message buffer in the stack,
as a result it may contain random data on each call to it, this is no
problem for PG >= 14, since the HASH_STRINGS flag used on hash table
creation ensures that only the text before the '\0' is used in the key.

For PG <= 13, there is not HASH_STRINGS flag, so it considers the entire
key data to calculate the hash, this poses a problem since we use the
error message as key and it may contain random data on each call.

To fix the problem, for PG <= 13 versions, after we add the message to
the error buffer, we fill the remaining bytes on it with zeroes, this
ensures key consistency for strings across different PG versions.